% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_contrasts1}
\alias{df_contrasts1}
\title{Simulated data with response times as a function of a between-subject 2-level factor F.}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{F}{Between subject factor with factor levels F1 and F2}
\item{DV}{Dependent variable, response time in seconds}
\item{id}{Subject index}
}
}
\usage{
df_contrasts1
}
\description{
The condition means are exactly 0.8 and 0.4 seconds.
}
\details{
The data were simulated using the following R-code:

set.seed(123)

d <- mvrnorm(n=5, mu=c(0.8, 0.4), Sigma=diag(2)*0.2^2, empirical=TRUE)

df_contrasts1 <- d %>% as.data.frame() %>% gather(key="F", value="DV") %>%
  mutate(id=1:nrow(.), F=factor(F))

levels(df_contrasts1$F) <- c("F1", "F2")
}
\references{
\insertAllCited{}
}
\keyword{datasets}
