% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_contrasts2}
\alias{df_contrasts2}
\title{Simulated data with response times from a lexical decision task as a function of a 3-level between-subject factor "word class".}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{F}{Between subject factor word class with factor levels "nouns", "verbs", and "adjectives"}
\item{DV}{Dependent variable, response time in milliseconds}
\item{id}{Subject index}
}
}
\usage{
df_contrasts2
}
\description{
The simulated data contains data from word classes "nouns", "verbs", and "adjectives". Word class is manipulated between subjects here (usually it is a within-subject manipulation).
The condition means are exactly 500, 450, and 400 milliseconds for nouns, verbs, and adjectives words.
}
\details{
The data were simulated using the following R-code:

set.seed(123)

d2 <- mvrnorm(n=4, mu=c(500, 450, 400),
              Sigma=diag(3)*20^2, empirical=TRUE)

df_contrasts2 <- d2 %>% as.data.frame() %>%
  gather(key="F", value="DV") %>%
  mutate(id=1:nrow(.), F=factor(F))

levels(df_contrasts2$F) <- c("nouns", "verbs", "adjectives")

df_contrasts2$DV <- round(df_contrasts2$DV)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
