% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_contrasts3}
\alias{df_contrasts3}
\title{Simulated data with response times as a function of a 4-level between-subject factor.}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{F}{Between subject factor with factor four levels (F1, F2, F3, F4)}
\item{DV}{Dependent variable}
\item{id}{Subject index}
}
}
\usage{
df_contrasts3
}
\description{
The condition means are exactly 10, 20, 10, and 40.
It could be assumed that the four factor levels `F1` to `F4` reflect levels of word frequency, including the levels `low`, `medium-low`, `medium-high`, and `high` frequency words, and that the dependent variable (DV) reflects some response time.
}
\details{
The data were simulated using the following R-code:

set.seed(123)

d3 <- mvrnorm(n=5, mu=c(10, 20, 10, 40), Sigma=diag(4)*10^2, empirical=TRUE)

df_contrasts3 <- d3 %>% as.data.frame() %>% gather(key="F", value="DV") %>%
  mutate(id=1:nrow(.), F=factor(F))

levels(df_contrasts3$F) <- c("F1", "F2", "F3", "F4")
}
\references{
\insertAllCited{}
}
\keyword{datasets}
