% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_contrasts4}
\alias{df_contrasts4}
\title{Simulated data with response times from a 2 x 2 between subject design.}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{A}{Between subject factor with two levels (A1, A2)}
\item{B}{Between subject factor with two levels (B1, B2)}
\item{DV}{Dependent variable}
\item{id}{Subject index}
}
}
\usage{
df_contrasts4
}
\description{
The condition means are exactly 10, 20, 10, and 40; means and standard deviations are exactly the same as in `df_contrasts3`.
}
\details{
The data were simulated using the following R-code:

set.seed(123)

d4 <- mvrnorm(n=5, mu=c(10, 20, 10, 40), Sigma=diag(4)*10^2, empirical=TRUE)

df_contrasts4 <- d4 %>% as.data.frame() %>% gather(key="A", value="DV") %>%
  mutate(id=1:nrow(.), B=factor(A), A=factor(A))

levels(df_contrasts4$A) <- c("A1","A1","A2","A2")

levels(df_contrasts4$B) <- c("B1","B2","B1","B2")
}
\references{
\insertAllCited{}
}
\keyword{datasets}
