% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_contrasts5}
\alias{df_contrasts5}
\title{Simulated data with response times with a 2-level between subject factor and IQ as covariate: no interaction.}
\format{
A data frame with 60 rows and 4 variables:
\describe{
\item{F}{Between subject factor with factor two levels (F1, F2)}
\item{RT}{Dependent variable response time (RT)}
\item{IQ}{Between-subject covariate (IQ)}
\item{id}{Subject index}
}
}
\usage{
df_contrasts5
}
\description{
In the data, the influence of IQ on response time is exactly the same in both conditions F1 and F2. However, the mean response time and mean IQ differs between conditions F1 and F2.
}
\details{
The data were simulated using the following R-code:

set.seed(123)

N <- 30

mu1 <- c(225, 85)

mu2 <- c(200,115)

sd1 <- 40

sd2 <- 15

r1 <- r2 <- -0.4

sigma1 <- matrix(c(sd1^2,r1*sd1*sd2,r1*sd1*sd2,sd2^2),nrow=2)

sigma2 <- matrix(c(sd1^2,r2*sd1*sd2,r2*sd1*sd2,sd2^2),nrow=2)

d5a <- mvrnorm(n=N, mu=mu1, Sigma=sigma1, empirical=TRUE)

d5b <- mvrnorm(n=N, mu=mu2, Sigma=sigma2, empirical=TRUE)

df_contrasts5 <- rbind(data.frame(F="F1", d5a), data.frame(F="F2", d5b))

names(df_contrasts5)[2:3] <- c("RT","IQ")

df_contrasts5$id <- 1:nrow(df_contrasts5)

df_contrasts5$F  <- factor(df_contrasts5$F)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
