% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_contrasts7}
\alias{df_contrasts7}
\title{Simulated data with a binomial response from a 2 x 2 between subject design.}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{A}{Between subject factor with two levels (A1, A2)}
\item{B}{Between subject factor with two levels (B1, B2)}
\item{pDV}{Dependent variable: successful task performance (0 = no success versus 1 = success)}
\item{id}{Subject index}
}
}
\usage{
df_contrasts7
}
\description{
The dependent variable could be successful task performance. The condition means are exactly 0.2, 0.5, 0.2, and 0.8.
}
\details{
The data were simulated using the following R-code:

set.seed(123)

N <- 50

d7 <- data.frame(A  =factor(rep(c("A1","A1","A2","A2"), each=N)),

                 B  =factor(rep(c("B1","B2","B1","B2"), each=N)),

                 pDV=       rep(c( 0.2, 0.5, 0.2, 0.8), each=N))

tmp <- c(seq(0,1,length=N)[sample(N)], seq(0,1,length=N)[sample(N)],

         seq(0,1,length=N)[sample(N)], seq(0,1,length=N)[sample(N)])

d7$pDV <- ifelse(d7$pDV > tmp, 1, 0)

d7$id <- 1:nrow(d7)

df_contrasts7 <- tibble(d7)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
