% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dillonE1}
\alias{df_dillonE1}
\title{Data from an eyetracking reading experiment that records total reading times in milliseconds at an auxiliary verb in so-called low and high interference conditions. \insertCite{DillonEtAl2013;textual}{lingpsych}}
\format{
A data frame with 2,855 rows and 5 variables:
\describe{
\item{subj}{The subject id, ranging from 1 to 40.}
\item{item}{The item id, ranging from 1 to 48. Each item id is prefixed by dillonE1.}
\item{rt}{The total reading times at the critical region (auxiliary verb) in milliseconds.}
\item{int}{The two levels of the interference manipulation, low and high.}
\item{expt}{The experiment id.}
}
}
\usage{
df_dillonE1
}
\description{
The dataset is from an eyetracking experiment by \insertCite{DillonEtAl2013;textual}{lingpsych}, and contains reading times (total reading times) from a design that orginally had eight conditions. There are 40 subjects who read 48 items in a Latin square design. The data shown here focus on two conditions from the experiment, high and low interference conditions in a subject-verb number agreement configuration in ungrammatical sentences.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
