% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dillonrep}
\alias{df_dillonrep}
\title{Reading times at the critical regions of an eyetracking experiment conducted by  \insertCite{JaegerMertzenVanDykeVasishth2019;textual}{lingpsych}.}
\format{
A data frame with 8,628 rows and 8 variables:
\describe{
 \item{subj}{Subject id.}
 \item{item}{Item id.}
\item{cond}{The conditions labels, ranging from a to h.}
\item{acc}{Question-response accuracy.}
\item{roi}{The region of interest.}
\item{FPRT}{First-pass reading time.}
\item{TFT}{Total fixation time.}
\item{FPR}{First-pass regression.}
}
}
\usage{
df_dillonrep
}
\description{
Reading times at the critical region of sentences of the experiment reported in \insertCite{JaegerMertzenVanDykeVasishth2019;textual}{lingpsych}.
}
\details{
The experiment, which is a replication attempt of Dillon et al. 2013, examined sentences with eight conditions. 

The data and associated code are available from https://osf.io/reavs/.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
