% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_discreteagrmt}
\alias{df_discreteagrmt}
\title{Data from an experiment involving an agreement attraction design with Likert ratings and accuracy of response \insertCite{AnnaLphd;textual}{bcogsci}}
\format{
A data frame with 4,296 rows and 7 variables:
\describe{
\item{item}{Item id.}
\item{subj}{Subject id.}
\item{attractor}{A factor with two levels, match and mismatch.}
\item{interference}{A factor with two levels, interference and no interference.}
\item{rating}{A Likert scale rating of the acceptability of the sentence.}
\item{rating}{A Likert scale rating of the acceptability of the sentence.}
\item{accuracy}{Question response accuracy.}
\item{Sentence}{The sentence that the subject was shown.}
}
}
\usage{
df_discreteagrmt
}
\description{
The dataset, from experiment 1 in chapter 4 of \insertCite{AnnaLphd;textual}{bcogsci}, contains the responses from 4,296 subjects who participated in an internet-based study that elicited Likert ratings and question-response accuracies for a single sentence they were shown. There were two levels of a factor called attractor (levels: match and mismatch) and two levels of another factor called interference (levels: interference and no interference).
}
\references{
\insertAllCited{}
}
\keyword{datasets}
