% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dutch}
\alias{df_dutch}
\title{Data from a self-paced reading experiment that records log reading times in log milliseconds of grammatical and ungrammatical double center embeddings in Dutch. \insertCite{FrankEtAl2015;textual}{lingpsych}}
\format{
A data frame with 372 rows and 4 variables:
\describe{
\item{subject}{The subject id, ranging from 1 to 24. Each subject id is preceded by an s.}
\item{item}{The item id, ranging from 1 to 16. Each subject id is preceded by an i.}
\item{condition}{Grammatical conditions are coded as +1, and ungrammatical conditions as -1.}
\item{NP1}{The log reading times at the noun phase following the final verb in a double center embedding.}
}
}
\usage{
df_dutch
}
\description{
The dataset is from a self-paced reading study by \insertCite{FrankEtAl2015;textual}{lingpsych}, and contains reading times from a design that had two conditions, grammatical and ungrammatical double center embeddings in Dutch. There are 24 subjects who read 16 items in a Latin square design. The data shown here focus on log reading times at the noun phrase following the final verb in a double center embedding.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
