% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_english}
\alias{df_english}
\title{Data from a self-paced reading experiment that records log reading times in log milliseconds of grammatical and ungrammatical double center embeddings in English. \insertCite{VSLK08;textual}{lingpsych}}
\format{
A data frame with 768 rows and 4 variables:
\describe{
\item{subject}{The subject id, ranging from 1 to 49.}
\item{item}{The item id, ranging from 1 to 16.}
\item{condition}{Grammatical conditions are coded as +1, and ungrammatical conditions as -1.}
\item{NP1}{The log reading times at the noun phase following the final verb in a double center embedding.}
}
}
\usage{
df_english
}
\description{
The dataset is from a self-paced reading study by \insertCite{VSLK08;textual}{lingpsych}, and contains reading times from a design that had two conditions, grammatical and ungrammatical double center embeddings in English. There are 48 subjects who read 16 items in a Latin square design. The data shown here focus on log reading times at the noun phrase following the final verb in a double center embedding.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
