% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_F1data}
\alias{df_F1data}
\title{Data from an experiment with F1 frequency in Hertz of vowels produced by male and females speakers of different languages \insertCite{johnson2011quantitative;textual}{lingpsych}}
\format{
A data frame with 19 rows and 4 variables:
\describe{
\item{female}{F1 frequency data in Hertz from females.}
\item{male}{F1 frequency data in Hertz from males.}
\item{vowel}{The vowel produced.}
\item{language}{The language spoken by the speaker.}
}
}
\usage{
df_F1data
}
\description{
The dataset is from \insertCite{johnson2011quantitative;textual}{lingpsych}, contains the F1 frequencies recorded from 19 male subjects and 19 female subjects who produced vowels in their native languages.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
