% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_fedorenko06}
\alias{df_fedorenko06}
\title{Reading times at the critical region () of a self-paced reading experiment conducted by  \insertCite{fedorenko06;textual}{lingpsych}.}
\format{
A data frame with 1,440 rows and 7 variables:
\describe{
 \item{subj}{Subject id.}
 \item{item}{Item id.}
\item{RT}{Reading times in milliseconds at the critical region.}
\item{rctype}{The two levels of a factor representing the relative clause type.}
\item{nountype}{The two levels of a factor representing the noun type (name or occupation).}
\item{load}{The two levels of a factor representing memory load (three words or one word).}
\item{region}{The critical region in the sentence.}
}
}
\usage{
df_fedorenko06
}
\description{
Reading times at the critical region of sentences of experiment 2 reported in \insertCite{fedorenko06;textual}{lingpsych}.
}
\details{
The experiment examined sentences with eight conditions (a 2x2x2 design). 
This self-paced experiment examined sentences with eight conditions (a 2x2x2 design). 
One factor is relative clause type (object or subject relative); 
another factor is memory load (three or one items held in memory); 
the third factor is nountype, which represents items held in memory that 
are either similar or dissimilar to the nouns in the sentences. 

#'
}
\references{
\insertAllCited{}
}
\keyword{datasets}
