% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_gg05e1}
\alias{df_gg05e1}
\title{Data from a self-paced reading experiment that records reading times in milliseconds of subject and object relative clauses in English. \insertCite{grodner;textual}{lingpsych}}
\format{
A data frame with 672 rows and 4 variables:
\describe{
\item{subject}{The subject id, ranging from 1 to 42.}
\item{item}{The item id, ranging from 1 to 16.}
\item{condition}{Condition labels; objgap represents object relatives, and subjgap represents subject relatives.}
\item{rawRT}{The reading times at the critical region (the relative clause verb) in milliseconds.}
}
}
\usage{
df_gg05e1
}
\description{
The dataset is from Experiment 1 in \insertCite{grodner;textual}{lingpsych}, and contains reading times from a self-paced reading experiment on subject and object relative clauses in English. There are 42 subjects who read 16 items in a Latin square design. The data are from the relative clause verb, which is the critical region of interest.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
