% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_gibsonwu}
\alias{df_gibsonwu}
\title{Data from a self-paced reading experiment that records reading times in milliseconds of subject and object relative clauses in Chinese. \insertCite{gibsonwu;textual}{lingpsych}}
\format{
A data frame with 547 rows and 4 variables:
\describe{
\item{subj}{The subject id, ranging from 1 to 40.}
\item{item}{The item id, ranging from 1 to 16.}
\item{type}{Condition labels; obj-ext represents object relatives, and subj-ext represents subject relatives.}
\item{rt}{The reading times at the critical region (the relative clause head noun) in milliseconds.}
}
}
\usage{
df_gibsonwu
}
\description{
The dataset is from an experiment reported in \insertCite{gibsonwu;textual}{lingpsych}, and contains reading times from a self-paced reading experiment on subject and object relative clauses in Chinese. There are 37 subjects who read 15 items in a Latin square design. The data are from the relative clause head noun, which is the critical region of interest.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
