% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_gibsonwu2}
\alias{df_gibsonwu2}
\title{Data from a self-paced reading experiment that records reading times in milliseconds of subject and object relative clauses in Chinese. \insertCite{VasishthetalPLoSOne2013;textual}{lingpsych}}
\format{
A data frame with 595 rows and 6 variables:
\describe{
\item{subj}{The subject id, ranging from 1 to 40.}
\item{item}{The item id, ranging from 1 to 16.}
\item{condition}{Condition labels; obj-ext represents object relatives, and subj-ext represents subject relatives.}
\item{pos}{The word position of the critical region in the sentence.}
\item{rt}{The reading times at the critical region (the relative clause head noun) in milliseconds.}
\item{region}{The label for the critical region (head noun) in the sentence.}
}
}
\usage{
df_gibsonwu2
}
\description{
The dataset is from a replication attempt of experiment by \insertCite{gibsonwu;textual}{lingpsych}; the replication attempt was reported in \insertCite{VasishthetalPLoSOne2013;textual}{lingpsych}, and contains reading times from a self-paced reading experiment on subject and object relative clauses in Chinese. There are 40 subjects who read 15 items in a Latin square design. The data are from the relative clause head noun, which is the critical region of interest.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
