% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_levykeller13E1}
\alias{df_levykeller13E1}
\title{Total reading times at the critical region (verb) of an eyetracking experiment conducted by  \insertCite{levy2013expectation}{lingpsych}.}
\format{
A data frame with 672 rows and 5 variables:
\describe{
 \item{subj}{Subject id.}
 \item{item}{Item id.}
\item{dat}{The two levels of a factor representing whether 
a dative phrase is in the main or subordinate clause.}
\item{adj}{The two levels of a factor representing whether an adjust is in the main or subordinate clause.}
\item{TFT}{Total reading times in milliseconds at the critical region.}
}
}
\usage{
df_levykeller13E1
}
\description{
Total reading times at the critical region of sentences of experiment 1 reported in \insertCite{levy2013expectation;textual}{lingpsych}.
}
\details{
The experiment examined sentences with four conditions (a 2x2 design). 
This eyetracking experiment examined sentences with four conditions (a 2x2 design). 
One factor is whether a dative phrase is in the main or subordinate clause;
another factor is whether an adjunct phrase is in the main or subordinate clause.

The conditions (simplified) are:
a. Nachdem der Lehrer [ADJ zur Ahndung] [DAT dem Sohn] ... hat Hans Gerntner ... den Fussball versteckt und somit...
b. Nachdem der Lehrer [               ] [DAT dem Sohn] ... hat Hans Gerntner ... [ADJ zur Ahndung] den Fussball versteckt...
c. Nachdem der Lehrer [ADJ zur Ahndung] [            ] ... hat Hans Gerntner ... [DAT dem Sohn] den Fussball versteckt...
d. Nachdem der Lehrer [               ] [            ] ... hat Hans Gerntner ... [ADJ zur Ahndung] [DAT dem Sohn] 
den Fussball versteckt...
}
\references{
\insertAllCited{}
}
\keyword{datasets}
