% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_polarity}
\alias{df_polarity}
\title{Reading times at the critical regions of an eyetracking experiment conducted by  \insertCite{VBLD07;textual}{lingpsych}.}
\format{
A data frame with 8,823 rows and 5 variables:
\describe{
 \item{subject}{Subject id, ranging from 1 to 48.}
 \item{item}{Item id, ranging from 1 to 36.}
\item{condition}{The conditions labels, ranging from a to f.}
\item{times}{The different dependent measures from the eyetracking study.}
\item{value}{The dependent measure (milliseconds).}
}
}
\usage{
df_polarity
}
\description{
Reading times at the critical region of sentences of the experiment reported in \insertCite{VBLD07;textual}{lingpsych}.
}
\details{
The experiment examined sentences with six conditions: three conditions with negative polarity items (one grammatical, two ungrammatical), 
and three with positive polarity items (two grammatical, one ungrammatical). 

The data and associated code are available from https://github.com/vasishth/ProcessingPolarity.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
