% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_VMJG18E7}
\alias{df_VMJG18E7}
\title{Reading times at the critical region (verb) of an eyetracking experiment conducted by  \insertCite{VasishthMertzenJaegerGelman2018}{lingpsych}.}
\format{
A data frame with 2,384 rows and 4 variables:
\describe{
 \item{subj}{Subject id.}
 \item{item}{Item id.}
\item{condition}{The four condition levels.}
\item{TFT}{Total reading times in milliseconds at the critical region.}
}
}
\usage{
df_VMJG18E7
}
\description{
Reading times at the critical region of sentences of experiment 7 reported in \insertCite{VasishthMertzenJaegerGelman2018;textual}{lingpsych}.
}
\details{
The experiment examined sentences with four conditions (a 2x2 design). 
This eyetracking experiment examined sentences with four conditions (a 2x2 design). 
One factor is whether a dative phrase is in the main or subordinate clause;
another factor is whether an adjunct phrase is in the main or subordinate clause.

The conditions (simplified) are:
a. Nachdem der Lehrer [ADJ zur Ahndung] [DAT dem Sohn] ... hat Hans Gerntner ... den Fussball versteckt und somit...
b. Nachdem der Lehrer [               ] [DAT dem Sohn] ... hat Hans Gerntner ... [ADJ zur Ahndung] den Fussball versteckt...
c. Nachdem der Lehrer [ADJ zur Ahndung] [            ] ... hat der Mitschuler, der [DAT dem Sohn] den Fussball versteckt hat, ...
d. Nachdem der Lehrer [               ] [            ] ... hat der Mitschuler, der [ADJ zur Ahndung] [DAT dem Sohn] 
den Fussball versteckt hat, ...
}
\references{
\insertAllCited{}
}
\keyword{datasets}
