% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dillonE1ttnested}
\alias{dillonE1ttnested}
\title{Model fit for an eyetracking reading experiment that records total reading times in milliseconds for a 2x2x2 design. \insertCite{DillonEtAl2013;textual}{lingpsych}}
\format{
A brms model fit
\describe{
}
}
\usage{
dillonE1ttnested
}
\description{
The dataset is a model fit for an English eyetracking reading study by \insertCite{DillonEtAl2013;textual}{lingpsych}. There are 40 subjects who read 48 items in a Latin square design. The data shown here focus on total reading times at the critical region.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
