% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_lnorm_functions.R
\name{gen_fake_lnorm2x2x2}
\alias{gen_fake_lnorm2x2x2}
\title{Generate lognormal data from an eight-condition repeated measures design; the function contains contrast coding for a particular experiment design by Dillon et al. 2013.}
\usage{
gen_fake_lnorm2x2x2(
  ncond = 8,
  nitem = NULL,
  nsubj = NULL,
  beta = NULL,
  Sigma_u = NULL,
  Sigma_w = NULL,
  sigma_e = NULL
)
}
\arguments{
\item{ncond}{Number of conditions (eight).}

\item{nitem}{Number of items, must be divisible by eight.}

\item{nsubj}{Number of subjects, must be divisible by eight.}

\item{beta}{The intercept and slopes of the fixed effects.}

\item{Sigma_u}{The variance-covariance matrix for subject random effects.}

\item{Sigma_w}{The variance-covariance matrix for item random effects.}

\item{sigma_e}{The standard deviation of the residuals.}
}
\value{
A data frame containing simulated data.
}
\description{
This function takes parameters from a previously fitted model 
and returns simulated data.
}
