% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_lnorm_functions.R
\name{gen_sim_lnorm2}
\alias{gen_sim_lnorm2}
\title{Generate lognormal data from a two-condition repeated measures design (Latin square)}
\usage{
gen_sim_lnorm2(
  nitem = 16,
  nsubj = 42,
  beta = NULL,
  Sigma_u = NULL,
  Sigma_w = NULL,
  sigma_e = NULL
)
}
\arguments{
\item{nitem}{Number of items, must be divisible by two.}

\item{nsubj}{Number of subjects, must be divisible by two.}

\item{beta}{The intercept and slope of the fixed effects.}

\item{Sigma_u}{The variance-covariance matrix for subject random effects.}

\item{Sigma_w}{The variance-covariance matrix for item random effects.}

\item{sigma_e}{The standard deviation of the residuals.}
}
\value{
A data frame containing simulated data.
}
\description{
This function takes parameters from a previously fitted model 
and returns simulated data.
}
